requirements.txt
jupyterlab>3.0
flask==2.0.3
python-dotenv==0.19.0
transformers==4.30.0
gemini-sdk==0.2.5

